+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/srs/ui/typeselection/SrsTypeSelectionFragmentViewModel.kt
@@ -1,4 +1,4 @@
-package de.rki.coronawarnapp.srs.ui.typeselction
+package de.rki.coronawarnapp.srs.ui.typeselection
 
 import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
