+++ Corona-Warn-App/src/main/res/layout/fragment_srs_submission_consent.xml
