+++ Corona-Warn-App/src/main/res/layout/include_srs_privacy_card.xml
