+++ Corona-Warn-App/src/main/res/values-de/srs_submission_strings.xml
