+++ Corona-Warn-App/src/main/res/values/srs_submission_strings.xml
